Added Text To Speech Command to TestClient.  When activated, client will echo all recieved chat messages out thru the computer's speakers.
Since this requires .Net 3.0 I've left it out of the project by default.  Nant files may need to be edited to exclude it.
To use this: include it in the project and add a reference to the System.Speech.dll

It's like turning off Close Captioning!  All y'all just gotta try it.

git-svn-id: svn://opensecondlife.org/libsl/trunk@797 c754d5e9-d413-0410-bcdf-9574c9dba6b7